5/20 Nightly
5/20 Nightly

+++ webdna.sh
@@ -1,29 +1,61 @@
 #!/bin/sh
 
-DNA_FOLDER="${DNA_FOLDER:WebDNA-folder-8.6}"
-DNA_EXECUTABLE="${DNA_EXECUTABLE:WebDNA-Linux-FastCGI-8.6.5}"
-WEBROOT="${WEBROOT:/var/www/html}"
-WEBDNA_LOC="${WEBDNA_LOC:$WEBROOT/WebDNA}"
-APACHE_USER="${APACHE_USER:apache}"
-APACHE_LOC="${APACHE_LOC:/etc/httpd}"
-APACHE_CONFIG="${APACHE_CONFIG:/etc/httpd/conf/httpd.conf}"
+DNA_FOLDER="${DNA_FOLDER:-WebDNA-folder-8.6}"
+DNA_EXECUTABLE="${DNA_EXECUTABLE:-WebDNA-Linux-FastCGI-8.6.5}"
+
+WEBROOT="${WEBROOT:-/var/www/html}"
+WEBDNA_LOC="${WEBDNA_LOC:-$WEBROOT/WebDNA}"
+APACHE_USER="${APACHE_USER:-apache}"
+APACHE_LOC="${APACHE_LOC:-/etc/httpd}"
+APACHE_CONFIG="${APACHE_CONFIG:-$APACHE_LOC/conf/httpd.conf}"
+
+PATTERN='WebDNA-(Linux|Alpine)-(FastCGI|Server)-([0-9].[0-9].[0-9])'
+
+GET_VALUE(){
+    [[ $1 =~ $PATTERN ]] && echo "${BASH_REMATCH[$2]}"
+}
+
+SET_VERSION(){ 
+    DNA_VERSION="$(GET_VALUE $DNA_EXECUTABLE 3)" 
+}
+
+SET_PLATFORM(){
+    DNA_PLATFORM="$(GET_VALUE $DNA_EXECUTABLE 1)"
+}
+
+SET_TYPE(){
+    DNA_TYPE="$(GET_VALUE $DNA_EXECUTABLE 2)"
+}
+
+# Set Stuff
+SET_VERSION
+SET_PLATFORM
+SET_TYPE
 
  # Pull in the WebDNA folder and fcgi
-wget http://webdna.us/download/${DNA_FOLDER}.zip
+wget https://webdna.us/download/${DNA_FOLDER}.zip
 unzip -d . "${DNA_FOLDER}.zip"
 rm -f "${DNA_FOLDER}.zip"
-wget http://www.webdna.us/download/${DNA_EXECUTABLE}.zip
+wget https://www.webdna.us/download/${DNA_EXECUTABLE}.zip
 unzip -d . "${DNA_EXECUTABLE}.zip"
 rm -f "${DNA_EXECUTABLE}.zip"
 
 # 'Install' WebDNA
-mv "${DNA_EXECUTABLE}/WebDNA.fcgi" "WebDNA/WebDNA.fcgi"
-chmod 0755 "WebDNA/WebDNA.fcgi"
-mv "WebDNA" ${WEBDNA_LOC}
-chown -R $APACHE_USER:$APACHE_USER ${WEBDNA_LOC}
+if [ "${DNA_TYPE}" == "FastCGI" ]; then
 
-# Remove WebDNA Artifacts
-rm -rf "${DNA_EXECUTABLE}"
+    if [ "${DNA_VERSION}" == "8.6.5" ]; then
+        mv "${DNA_EXECUTABLE}/WebDNA.fcgi" "WebDNA/WebDNA.fcgi";
+        # Remove WebDNA Artifacts
+        rm -rf "${DNA_EXECUTABLE}";
+    else
+        mv "WebDNA.fcgi" "WebDNA/WebDNA.fcgi";
+    fi
+    chmod 0755 "WebDNA/WebDNA.fcgi"
+    mv "WebDNA" ${WEBDNA_LOC}
+
+fi
+
+chown -R $APACHE_USER:$APACHE_USER ${WEBDNA_LOC}
 
 # Quick Config of Apache
 mkdir -p "$WEBROOT/public"
